requirements.txt
fastapi
uvicorn
loguru
beautifulsoup4
requests
python-dotenv